Move content of config_reg_v6.h into epd_board_v6.c
